correction lab1

+++ src/Lab1/Ex2.java
@@ -2,7 +2,7 @@ package Lab1;
 
 public class Ex2 {
     public void calculateExpression (){
-        double result1 = (46 + 10) * (10.0 / 3);
+        double result1 = (46 + 10) * (double)10/ (double)3;
         double result2 = (29) * (4) * (-15);
 
         System.out.println("Result 1:"+ result1);
